requirements.txt
Django==4.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-resized==1.0.0
django-models==0.5.1
djangorestframework-simplejwt==5.2.2
djangorestframework==3.13.1
django-filter==22.1
decouple
requests
pillow
python-dotenv
python-whois
tqdm
boto3==1.26.41
botocore==1.29.41
django-storages==1.12.3


